+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Report type mismatches in `const` declarations
 - [Implement using function calls as values](https://github.com/amp-lang/ampc/issues/8)
 - [Implement implicit returns](https://github.com/amp-lang/ampc/issues/6)
+- Report diagnostic instead of panicking for mismatched return types
 
 ### [`snapshot-02-06-2023-a`](https://github.com/amp-lang/ampc/releases/tag/snapshot-02-06-2023-a)
 (initial snapshot)